+++ Cargo.toml
@@ -1,50 +1,15 @@
-[package]
-name = "typship"
+[workspace]
+resolver = "3"
+members = ["typship-cli", "typship-pack"]
+default-members = ["typship-cli"]
+
+[workspace.package]
 version = "0.4.2"
-description = "A cli for typst packages."
+edition = "2024"
 repository = "https://github.com/sjfhsjfh/typship"
 homepage = "https://github.com/sjfhsjfh/typship"
 readme = "README.md"
-edition = "2021"
 license-file = "LICENSE"
-
-[features]
-default = []
-
-[dependencies]
-anyhow = "1.0.95"
-clap = { version = "4.5.23", features = ["derive"] }
-crossterm = { version = "0.28.1", default-features = false }
-dialoguer = "0.11.0"
-dirs = "5.0.1"
-env_logger = "0.11.6"
-futures-util = "0.3.31"
-glob = "0.3.2"
-ignore = "0.4.23"
-log = "0.4.22"
-octocrab = { version = "0.42.1", features = [
-    "stream",
-    "default-client",
-    "rustls",
-    "rustls-ring",
-], default-features = false }
-regex = { version = "1.11.1", features = ["unicode-perl"], default-features = false }
-secrecy = { version = "0.10.3", features = ["serde"] }
-serde = { version = "1.0.216", features = ["derive"] }
-sha2 = "0.10.8"
-tempfile = "3.20.0"
-tokio = { version = "1.42.0", features = [
-    "rt",
-    "rt-multi-thread",
-    "macros",
-    "fs",
-    "io-util",
-    "net",
-], default-features = false }
-toml = "0.8.19"
-typst-syntax = "0.12.0"
-url = "2.5.4"
-whoami = "1.5.2"
 
 # The profile that 'dist' will build with
 [profile.dist]
