+++ typship-cli/src/commands/clean.rs
