+++ typship-cli/src/model.rs
