+++ typship-cli/src/model/manifest.rs
