+++ typship-cli/src/regs/mod.rs
